requirements.txt
feedparser==6.0.10
youtube-transcript-api>=1.2.3
python-dateutil==2.8.2
requests==2.31.0
docling>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

